Refactor : 불필요 이미지 파일 정리, 코드 정리
