--- src/components/User/SnsLogin/SnsLoginStyle.jsx
@@ -1,56 +1,0 @@
-import styled from "styled-components";
-
-export const SelectTabBar = styled.div`
-  display: flex;
-  margin-top: 28px;
-  width: 100%;
-  margin-bottom: 28px;
-`;
-
-export const SelectTabItem = styled.div`
-  flex: 1 1;
-  font-size: ${(props) => props.theme.fontSizeH4};
-  font-weight: 400;
-  font-size: 18px;
-  line-height: 26px;
-  text-align: center;
-  cursor: pointer;
-  font-weight: ${(props) => (props.isActive ? 700 : 400)};
-  color: ${(props) => (props.isActive ? "#0066FF" : "#888888")};
-  a {
-    display: block;
-  }
-`;
-
-export const SnsLinkList = styled.ul`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
-
-export const SnsLinkItem = styled.li`
-  width: 100%;
-`;
-
-export const LinkItem = styled.div`
-  height: 60px;
-  margin-bottom: 20px;
-  border-radius: 12px;
-  border: 1px solid #cccccc;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  /* img {
-    width: 100%;
-    height: 70px;
-  } */
-`;
-
-export const RowGroup = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-evenly;
-  margin-top: 32px;
-`;
