+++ src/pages/NotFoundScene.jsx
@@ -1,15 +1,9 @@
 import { useLocation, Link } from "react-router-dom";
+import styled from "styled-components";
 import Title from "../utils/Title/Title";
 import Nav from "../components/Navbar/index";
 import { WidthAreaSpace } from "../styles/Layout/Layout";
 import { FileBtn } from "../components/Board/NewWrite/BoardWriteStyle";
-import styled from "styled-components";
-
-const WrongDiv = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
 
 function NotFoundScene() {
   const { pathname } = useLocation();
@@ -29,5 +23,11 @@ function NotFoundScene() {
     </>
   );
 }
+
+const WrongDiv = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
 
 export default NotFoundScene;
