+++ src/pages/Profile.jsx
@@ -1,6 +1,6 @@
 import Title from "../utils/Title/Title";
 import Nav from "../components/Navbar/index";
-import ProfileContnet from "../components/Profile/Index";
+import ProfileContnet from "../components/Profile";
 import { WidthAreaSpace, WidthAreaInner } from "../styles/Layout/Layout";
 
 function Profile() {
